+++ the_tavern/models.py
@@ -208,6 +208,8 @@ class Survey(models.Model):
     waitlist_threshold = models.IntegerField(null=True, blank=True)
     has_waitlist = models.BooleanField(default=False)
     is_registration = models.BooleanField(default=False, help_text='Display tournament rules with a required agreement checkbox.')
+    auto_enroll = models.BooleanField(default=False, help_text='Automatically add respondents to the linked series.')
+
 
     created_by = models.ForeignKey(Profile, on_delete=models.SET_NULL, null=True, blank=True, related_name='created_surveys')