+++ changelog.md
@@ -20,6 +20,7 @@
 - Better links on admin messages
 - Added rules link to registration surveys
 - Surveys check temporary guild membership when user takes survey
+- Option to auto enroll registration responses to Series
 
 ### Bug Fixes
 - allow other option populates correctly when loading from template